qwerty

+++ app/src/main/java/uz/texnopos/installment/data/model/Payment.kt
@@ -1,6 +1,6 @@
 package uz.texnopos.installment.data.model
 
 data class Payment(
-    val order_id:Long,
+    val order_id:Int,
     val amount:Long
 )
